does this work

+++ examples/polybench_hls/adi.py
@@ -85,7 +85,7 @@ def adi(ttype, TSTEPS, N):
                 p[i, 0] = 0
                 q[i, 0] = v[0, i]
                 for j in range(1, N - 1):
-                    p[i, j] = (int) (-c / (a * p[i, j - 1] + b))
+                    p[i, j] = (int) (-c / ((int) (a * p[i, j - 1] + b)))
                     q[i, j] = (int) ((
                         -d * u[j, i - 1]
                         + (1 + 2 * d) * u[j, i]
